Create html, js, and computer choice function
Created html file, JavaScript file, and linked the JavaScript to the
html. I also wrote an initial function in the JavaScript file that
randomly returns the computer's action. Pseudocode was first written
in the form of comments, then it was fleshed out in JavaScript. I am
trying to put into practice what I learned regarding planning and
pseudocode.