Write a function that plays a round of the game
I wrote a function that plays a round of "rock", "paper", "scissors".
The function takes in the computer choice and the user choice. I
additionally wrote some code to test this and the computer choice
function, to simulate a round. Finally, I updated the computer choice
function to return a lower case value, and to print the choice of the
computer to the console. In a future commit I would like to rework the
conditional statements to make them more efficient.