Create game function for playing five rounds
I created a game function to play the game five times and record the
score in order to declare a winner. I also modified the play game
function in order to have it return a number for a switch statement
found in the new function. I now want to try and make the script more
efficient by simplifying the functions using helper functions.