Add event listeners to new html buttons
I have now begun to create a user interface for this rock, paper,
scissors game. I added event listeners to the newly made UI buttons,
and I removed the for loop found in the game function. Every other
re-work in noted as comments with a 2. Additionally, a new function was
created for the new game button to clear the board and start a new game
when clicked.