Improve upon UI
Removed the new game button, because it is not necessary for the time
being and for the scope of this project. The buttons representing the
moves disable after the game has been won. A confirm window appears
after the game has been won asking to continue the game or not. The
board is only cleared when a new game is played.