Organize index file into id names and class names
I made a design for the style of the webpage, and now I am getting the
site ready to be styled in this way. The html was edited to contain
four different boxes for css styling. These html sections were also
changed to have id names as the containing elements, and class names
as the inner elements. The javascript file was edited as well to reflect
these changes. Finally, a css file was created, but not yet populated.