Make the type writer function more efficient
I reduced the number of functions required to create the type-writer
effect to one, and therefor I only need one element to display the
results text. I also created a blinking cursor effect at the end of the
results text, however, I wasn't able to keep the cursor in-line with
the text. I also removed the pop-up containing the final results of the
game and the option of playing another game, because this feature was
becoming more and more complicated and out of the scope of this project.