Clean javascript code in script.js file
I removed the typeResults function as it's redundant, because all it
did was call teh type writer function. I also got rid of other the
other redundant typeWriter functions, because now there is only one
element to display all of the game results in the UI. I removed the
clear board function because I am no longer giving the option to restart
the game. I capitalized the final result text to make it more obvious to
the player, and I removed the option of a final tied game because this
option cannot occur. I updated the names of helper functions 1 and 2,
because they no longer print to the console, but assign strings to
the appropriate variables. Finally, I updated the comments based on
changes made.