Add final touches to the game
Added a hover effect in css (when the mouse cursor passes over the
button). In the html file, I removed the redundant elements found in the
text element. I added a setTimeout in the javascript that disables all
buttons after being clicked for a set period of time (700 ms). I added
a disabled effect on the buttons in css: when the buttons are disabled,
there is now no change in appearance. I added a favicon to the
html file, and changed the title to "Prehistory". I changed rock, paper,
and scissors to mountain, tree, and spears in the css classes, the html,
and in the results that are displayed as part of the dynamic javascript.
Finally, I changed the "make a move" message to "first one to five
wins", in order to give better context to the player.